requirements.txt
numpy>=1.24
scipy>=1.10
PyYAML>=6.0
pandas>=2.0
openpyxl>=3.1
joblib>=1.3
matplotlib>=3.7
gurobipy>=10.0
